feat(test): Remove unnecessary file tests and modify captcha test files
